+++ src/org/helioviewer/jhv/opengl/GLRenderer.java
@@ -36,6 +36,8 @@ public final class GLRenderer {
             case Latitudinal -> createConstantScales(viewports, MapScale.lati);
             case LogPolar -> createConstantScales(viewports, MapScale.logpolar(ImageLayers.getLargestRadialSize()));
             case Polar -> createConstantScales(viewports, MapScale.polar(ImageLayers.getLargestRadialSize()));
+            case LogDisk -> createConstantScales(viewports, MapScale.diskLog(ImageLayers.getLargestRadialSize()));
+            case PowerDisk -> createConstantScales(viewports, MapScale.diskPower(ImageLayers.getLargestRadialSize()));
         };
     }